Move data into src
