+++ src/components/GamePage/GamePage.jsx
@@ -7,11 +7,13 @@ import iconPerson from "../../assets/Icon-person.svg";
 import iconGamepad from "../../assets/Icon-gamepad.svg";
 import iconCross from "../../assets/Icon-cross.svg";
 import iconTimer from "../../assets/Icon-timer.svg";
+import GetRandomWord from "./GetRandomWord";
 
 export default function GamePage() {
   const userName = getDataFromLocalStorage('username');
   const difficultyLevel= getDataFromLocalStorage('difficultyLevel');
-  return (
+
+   return (
     <div className="gameContainer">
        <div className="gameHeader">
          <div className="playerDetails">
